Remove connector, basic librification
This updates the (dex) OIDC server code to remove the concept of
multiple connectors, and instead make it slightly better for consuming
within other services.

The external API is broken down in to two parts. First, is an auth
handler that wraps an existing handler. This is called when the client
starts the oauth flow, and will create a request entry in the store and
put it in the request context. The wrapped handler is then invoked,
which can then kick off whatever auth flow it wants

Secondly is the Finalize method. This can be called with the original
authorization request and the desired identity to provide to the user.
This will return a URL which when redirected to completes the client
OIDC flow.