Wire up server to app
This adjusts the server method to be embeddable externally (fixes #13)
and allows mounting it on an existing mux, with a configured auth
handler (fixes #14)

This is sketched in to app in a way that should give some indications
how it's intended to be used.

The interfaces/api is pretty janky right now (e.g passing server and
storage stuff), we can massage this later.