Implement acr_values and amr handling (#39)
* Implement acr_values and amr handling

This implements as per proposal in #35

the acr_values field is handled if requested, and passed through to the
connector. The connector can then choose to act or ignore these values.

The LoginRequest returned from the connector has a field for AMR and ACR. If
these are set, the values will be embedded in the returned token.

* Be more explicit about auth/refresh results

rather than implicitly overriding, make consumers of the mock connector pass
a function to generate the results.