Re-work claims
Rather than a simple map setup for claims, create a proper IDToken struct. This
represents the OIDC claims in a more... structured way. Implement appropriate
JSON handling for the OIDC types, as well as the ability to
serialize/deserialize extra fields.

Update the FinishAuthorization flow to _not_ take claims anymore - there is
somewhat of an overlap with this, and the metadata the implementor will pass.
Instead, make each call to the Token handler generate a full set of claims.
This makes the intent a bit more obvious that each call to that endpoint may
generate a different ID token, e.g if information changes between refresh
calls.