Coalesce the various token types into a session
It was getting kind of messy dealing with all the different stored token types,
their association, and where the metadata should live. Fix this by combining it
all into a "session", which represents the entire lifecycle from issuing a
code, through the access token and any refresh tokens. This makes tracking it
all far simpler.

To support this, a new issued token format was added. Rather than being a dot
separated string, it's now a base64 proto message. This gives us a better
format for the additional data we want to track, as well as making it more
extensible.