Move to more behavioural session interface
Previously we were using a generic storage interface, but after playing around
a bit it felt like this was hiding too much of the implementation. It forced
the user to track state via that as well, which became obviously more painful
when trying to do things like render all sessions a user currently has.

As this is meant to be a library, get a bit more focused and expose a simple
behavioural interface for tracking sessions. This plays nicely now that we've
settled on a "session" as the representation of the entire lifecycle. Users are
exposed the session ID, which they can use to manage state as they see fit. The
interface is a simple get/put/update, that can be backed on to whatever the
consumer feels like.

The session data (proto message) is hidden behind an interface, to avoid
building an implicit contract with the user.