Core OIDC Server (#3)
* Sketch out a really simple package and test case

Test is a simple end to end table rig, we can use for smoke testing the package
basics.

* Oauth2 server helpers

This adds a set of helper functions for handling the basic oauth2 auth and
token endpoint implementations. These do not implement the flow, but make it
easier to parse requests, send responses, and handle errors.

This has been kept private for now, as it's in flux.

* OIDC Flow implementation

This builds on the basic oauth2 helpers to implement an OIDC flow. Most of the
actual logic is left up to the consumer, we just abstract away the basic state
keeping and request/response processing.

* Token generation and management

We want to be able to issue opaque tokens that we can persist in a safe manner,
so that data leaks do not contain usable data.

This is implemented by generating tokens as an ID.TOKEN part. The ID is always
clear, and can be used to locate items in the database. The token part is
always stored using bcrypt.

* Add base demo IDP

We can use this to get a feel for actually implementing things, and as a
starting point for the conformance tests to target.

* Add sample RP

We can use this to exercise the OP. Super bare-bones for now.

* Re-work claims

Rather than a simple map setup for claims, create a proper IDToken struct. This
represents the OIDC claims in a more... structured way. Implement appropriate
JSON handling for the OIDC types, as well as the ability to
serialize/deserialize extra fields.

Update the FinishAuthorization flow to _not_ take claims anymore - there is
somewhat of an overlap with this, and the metadata the implementor will pass.
Instead, make each call to the Token handler generate a full set of claims.
This makes the intent a bit more obvious that each call to that endpoint may
generate a different ID token, e.g if information changes between refresh
calls.

* Ensure only the original client can redeem the code

* Add support for refresh_token grant to parser

* Coalesce the various token types into a session

It was getting kind of messy dealing with all the different stored token types,
their association, and where the metadata should live. Fix this by combining it
all into a "session", which represents the entire lifecycle from issuing a
code, through the access token and any refresh tokens. This makes tracking it
all far simpler.

To support this, a new issued token format was added. Rather than being a dot
separated string, it's now a base64 proto message. This gives us a better
format for the additional data we want to track, as well as making it more
extensible.

* Actually sign tokens

Actually use the configured signer, rather than using a static string for
tokens.

Move the shared test helpers into their own file.

* Re-work expiration a little

Rather than using fixed expiration times, allow the response to a token
request to set the expiry as desired.

Pass a little more info through the request/response too.

* Separate loading token endpoints with code and refresh

Use two separate parsing functions for this grant type. The different functions
are responsible for loading the appropriate token, and returning the
corresponding session.

* Move to more behavioural session interface

Previously we were using a generic storage interface, but after playing around
a bit it felt like this was hiding too much of the implementation. It forced
the user to track state via that as well, which became obviously more painful
when trying to do things like render all sessions a user currently has.

As this is meant to be a library, get a bit more focused and expose a simple
behavioural interface for tracking sessions. This plays nicely now that we've
settled on a "session" as the representation of the entire lifecycle. Users are
exposed the session ID, which they can use to manage state as they see fit. The
interface is a simple get/put/update, that can be backed on to whatever the
consumer feels like.

The session data (proto message) is hidden behind an interface, to avoid
building an implicit contract with the user.

* Drop example proto, no longer needed

* Implement refreshing

This adds actual issuance of refresh tokens.

* Remove token type & ID fields

These aren't really useful to us. The type can be inferred by the slot it was
stored in, and the ID was only useful when we stored multiple tokens in each
slot.  So drop them for now.

* Cleanup the proto a little

* Request refresh token in e2e test

* Implement Userinfo endpoint

This verifies that we have received a valid bearer token, and will return the
JSON information generated by a handler from the SessionID it's passed

Some more error code is added for the bearer token error responses.

* Handle ACR/AMR

Introduce a new "Authorization" type, that carries more information from
finishing the authorization flow through to the other endpoints.

Add parsing for the acr_values request

* Helper for filling ID tokens

Rather than leaving the user completely on their own, add a helper function to
the token request that pre-fills the ID token with the required, and the
informtion we know about.

* Update example server

We've extended functionality a bit, so update the example server to work. Also
add the ability to set some actual data for the returned token.

* Move to absolute expiry times

These are more flexible than durations.

* Remove superflous form parsing calls

* Remove overly strict message