OIDC Discovery support (#8)
* Add type for OIDC discovery metadata

This matches the JSON format served at the discovery endpoint

* Handler for provider metadata

This is a HTTP handler, that can take an optional keysource. It serves the main
metadata as JSON, and if they keysource is set it will also serve the JWKS URI.

* Add client for discovery endpoint

This can be passed an issuer, and used to inspect the metadata and fetch keys.

* Add valdiation for ProviderMetadata, and defaults

Simple validation to make sure the passed metadata is valid. Also add a handler
opt that will set some required fields based on what core supports.

* Add discovery support to example

This publishes the correct info on the op, and looks it up to configure the rp.

A fix to the KeySource interface was made so make it align with the Signer.

* Align client interface with Handler

No reason for it to differ so subtly

* Fix spelling