core: Make session storage easier to implement (#11)
We previously relied on the session manager to return the saved item as an
interface. In practice this is hard to implement, as the implementation has no
awareness of the concrete type that should be serialized/deserialized

Fix this by moving to an "into" method, ala proto/JSON marshaling. This allows
the code which is aware of the concrete type to instantiate it, and delegates
just the deserialization to the implementation.

This will make versioning slightly harder, as it's unclear what concrete type
we should be creating if there are multiple possibilities. I think this is OK
for now, in future we could do a two-pass setup to try and inspect a version
field (which we have reserved in the proto definition), or start using a
higher-level message that is just metadata and a google.protobuf.Any.