Allow token handlers to signal unauthorized requests
Calls to issue/refresh a token may be presented with a valid session, but the
handler might decide to deny it anyway.

Make this possible by checking the returned error from the handler to see if it
indicates the request failed because it was unauthorized, and return an
appropriate response to the caller.