Remove old proto storage, massage generation, promote proto to v1
Remove the old protobuf based storage, it's not been used for a while so it can
go to clean stuff up. Now only the versioned JSON is supported.

Update the proto build to use tools.go/shimmed local script, to make sure we
generate code with the expected version. This updates us to the latest
old-style proto package.

Promote the persisted types to v1. We don't use Any anywhere, so the package
name isn't a problem.