Update dependencies (#52)
* Update module to go 1.21, go-jose to v3

* Always need a valid code now, so default to internal service error

* Remove old proto storage, massage generation, promote proto to v1

Remove the old protobuf based storage, it's not been used for a while so it can
go to clean stuff up. Now only the versioned JSON is supported.

Update the proto build to use tools.go/shimmed local script, to make sure we
generate code with the expected version. This updates us to the latest
old-style proto package.

Promote the persisted types to v1. We don't use Any anywhere, so the package
name isn't a problem.

* Update to new protobuf package

* Update all modules, use built-in homedir func

* Correct lint/deprecation issues

* Switch to revive

* use built-in cache dir for default

rather than cluttering up HOME

* Update CI to always use latest stable go

* Use lint action